+++ node/src/types/kernel/api/filetree/removeDoc/payload.d.ts
@@ -15,13 +15,17 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-
-/* directories */
-export * as createDocWithMd from "./createDocWithMd";
-export * as removeDoc from "./removeDoc";
-export * as renameDoc from "./renameDoc";
-
-/* files */
-export * as listDocsByPath from "./listDocsByPath";
-export * as renameDoc from "./renameDoc";
-export * as searchDocs from "./searchDocs";
+/**
+ * Remove a document
+ */
+export interface IPayload {
+    /**
+     * notebook ID
+     */
+    notebook: string;
+    /**
+     * Document path, which needs to start with / and separate levels with /
+     * path here corresponds to the database path field
+     */
+    path: string;
+}