+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - 添加 API `/api/filetree/createDocWithMd` | Add API `/api/filetree/createDocWithMd`
 - 添加 API `/api/filetree/renameDoc` | Add API `/api/filetree/renameDoc`
+- 添加 API `/api/filetree/removeDoc` | Add API `/api/filetree/removeDoc`
 
 ## v0.0.5 / 2023-06-30